requirements.txt
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.0
pypdf>=4.0.0
docx2txt>=0.8
streamlit>=1.30.0
